Make it possible to keep the process buffer if it is visible.
Extend the setting of `prodigy-kill-process-buffer-on-stop` (and
related service option) with `'unless-visible` option to keep the
buffer alive if it is visible.

The workflow that inspired the change is something like this: have a
frame with all the log buffers visible... when we open the control
panel and do a restart everything stays in place and I don't have to
reopen all the buffers.

However, if some process runs in background and I stop it I would not
want to have hanging buffers around so they get cleaned automatically.