Make it possible to control service from its log buffer.
We extend the `prodigy-relevant-services` to return the "active"
service if the selected window is the log window.  This makes it
seamlessly work with almost all existing functions.

We introduce `prodigy-current-service` and use it instead of
`prodigy-service-at-pos` when it makes sense (jump-to-magit/directory
and similar).

Finally, bind `C-c` prefix in the view buffer to forward to
`prodigy-mode-map`---this way all the bindings will stay consistent
between the list and view buffers.